abstract repetitive code
