+++ hackathon_site/dashboard/frontend/src/components/dashboard/ItemTable/ItemTable.tsx
@@ -43,6 +43,8 @@ import {
     GeneralReturnTable,
 } from "components/general/OrderTables/OrderTables";
 import PopupModal from "components/general/PopupModal/PopupModal";
+import { sortPendingOrders, sortReturnedOrders } from "../../../api/helpers";
+import { sortCheckedOutOrders } from "../../../api/helpers";
 
 export const CheckedOutTables = () =>
     // TODO: for incident reports
@@ -51,14 +53,7 @@ export const CheckedOutTables = () =>
     {
         const dispatch = useDispatch();
         const unsorted_orders = useSelector(checkedOutOrdersSelector);
-        const orders = unsorted_orders.slice();
-
-        orders.sort((order1, order2) => {
-            return (
-                new Date(order2.updatedTime).valueOf() -
-                new Date(order1.updatedTime).valueOf()
-            );
-        });
+        const orders = unsorted_orders.slice().sort(sortCheckedOutOrders);
         const hardware = useSelector(hardwareSelectors.selectEntities);
         const isVisible = useSelector(isCheckedOutTableVisibleSelector);
         const fetchOrdersError = useSelector(orderErrorSelector);
@@ -216,83 +211,7 @@ export const CheckedOutTables = () =>
 export const ReturnedTable = () => {
     const dispatch = useDispatch();
     const unsorted_orders = useSelector(returnedOrdersSelector);
-    const orders = unsorted_orders.slice().sort((order1, order2) => {
-        const orderDate1 = order1.hardwareInOrder[0].time;
-        const orderDate2 = order2.hardwareInOrder[0].time;
-
-        const matchResult = orderDate1.match(
-            /(\d{1,2}):(\d{2}):(\d{2}) (AM|PM) \((\w{3}) (\w{3}) (\d{2}) (\d{4})\)/
-        );
-        const matchResult2 = orderDate2.match(
-            /(\d{1,2}):(\d{2}):(\d{2}) (AM|PM) \((\w{3}) (\w{3}) (\d{2}) (\d{4})\)/
-        );
-
-        if (matchResult && matchResult2) {
-            // converting invalid date to a valid Date for first order to be compared
-            const [, hours, minutes, seconds, meridiem, , month, day, year] =
-                matchResult;
-            const monthIndex = [
-                "Jan",
-                "Feb",
-                "Mar",
-                "Apr",
-                "May",
-                "Jun",
-                "Jul",
-                "Aug",
-                "Sep",
-                "Oct",
-                "Nov",
-                "Dec",
-            ].indexOf(month);
-
-            const date = new Date(
-                Number(year),
-                monthIndex,
-                Number(day),
-                Number(hours) + (meridiem === "PM" ? 12 : 0),
-                Number(minutes),
-                Number(seconds)
-            );
-            const formattedDate = date.toISOString();
-
-            // converting invalid date to a valid Date for first order to be compared
-            const [, hours2, minutes2, seconds2, meridiem2, , month2, day2, year2] =
-                matchResult2;
-            const monthIndex2 = [
-                "Jan",
-                "Feb",
-                "Mar",
-                "Apr",
-                "May",
-                "Jun",
-                "Jul",
-                "Aug",
-                "Sep",
-                "Oct",
-                "Nov",
-                "Dec",
-            ].indexOf(month2);
-
-            const date2 = new Date(
-                Number(year2),
-                monthIndex2,
-                Number(day2),
-                Number(hours2) + (meridiem2 === "PM" ? 12 : 0),
-                Number(minutes2),
-                Number(seconds2)
-            );
-            const formattedDate2 = date2.toISOString();
-
-            return (
-                new Date(formattedDate2).valueOf() - new Date(formattedDate).valueOf()
-            );
-        } else {
-            console.log("Invalid time format");
-        }
-        return 0;
-    });
-
+    const orders = unsorted_orders.slice().sort(sortReturnedOrders);
     const fetchOrdersError = useSelector(orderErrorSelector);
     const isVisible = useSelector(isReturnedTableVisibleSelector);
     const toggleVisibility = () => dispatch(toggleReturnedTable());
@@ -311,32 +230,9 @@ export const ReturnedTable = () => {
 
 export const PendingTables = () => {
     const dispatch = useDispatch();
-    const orders = useSelector(pendingOrderSelectors.selectAll);
-    let ready_orders = [];
-    let submitted_orders = [];
-    for (let order of orders) {
-        if (order.status === "Ready for Pickup") {
-            ready_orders.push(order);
-        } else {
-            submitted_orders.push(order);
-        }
-    }
-    ready_orders.sort((order1, order2) => {
-        return (
-            new Date(order1.updatedTime).valueOf() -
-            new Date(order2.updatedTime).valueOf()
-        );
-    });
-
-    submitted_orders.sort((order1, order2) => {
-        return (
-            new Date(order1.updatedTime).valueOf() -
-            new Date(order2.updatedTime).valueOf()
-        );
-    });
-
-    orders.splice(0, orders.length, ...ready_orders, ...submitted_orders);
-
+    const unsorted_orders = useSelector(pendingOrderSelectors.selectAll);
+    const orders = sortPendingOrders(unsorted_orders);
+    orders.reverse();
     const isVisible = useSelector(isPendingTableVisibleSelector);
     const isCancelOrderLoading = useSelector(cancelOrderLoadingSelector);
     const toggleVisibility = () => dispatch(togglePendingTable());
