+++ hackathon_site/dashboard/frontend/src/components/teamDetail/SimpleOrderTables/SimpleOrderTables.tsx
@@ -18,6 +18,7 @@ import {
     UpdateOrderAttributes,
     updateOrderStatus,
 } from "slices/order/teamOrderSlice";
+import { sortPendingOrders, sortReturnedOrders } from "../../../api/helpers";
 
 interface SimpleOrderFormValues {
     itemIdsChecked: string[];
@@ -142,32 +143,8 @@ const CompleteOrderButton = ({ order }: { order: OrderInTable }) => {
 
 export const SimplePendingOrderFulfillmentTable = () => {
     const dispatch = useDispatch();
-    const orders = useSelector(pendingOrdersSelector);
-    let ready_orders = [];
-    let submitted_orders = [];
-    for (let order of orders) {
-        if (order.status === "Ready for Pickup") {
-            ready_orders.push(order);
-        } else {
-            submitted_orders.push(order);
-        }
-    }
-    ready_orders.sort((order1, order2) => {
-        return (
-            new Date(order1.updatedTime).valueOf() -
-            new Date(order2.updatedTime).valueOf()
-        );
-    });
-
-    submitted_orders.sort((order1, order2) => {
-        return (
-            new Date(order1.updatedTime).valueOf() -
-            new Date(order2.updatedTime).valueOf()
-        );
-    });
-
-    orders.splice(0, orders.length, ...submitted_orders, ...ready_orders);
-
+    const unsorted_orders = useSelector(pendingOrdersSelector);
+    const orders = sortPendingOrders(unsorted_orders);
     const [isVisible, setVisibility] = useState(true);
     const toggleVisibility = () => setVisibility(!isVisible);
 
@@ -279,82 +256,7 @@ export const SimplePendingOrderFulfillmentTable = () => {
 
 export const AdminReturnedItemsTable = () => {
     const unsorted_orders = useSelector(returnedOrdersSelector);
-    const orders = unsorted_orders.slice().sort((order1, order2) => {
-        const orderDate1 = order1.hardwareInOrder[0].time;
-        const orderDate2 = order2.hardwareInOrder[0].time;
-
-        const matchResult = orderDate1.match(
-            /(\d{1,2}):(\d{2}):(\d{2}) (AM|PM) \((\w{3}) (\w{3}) (\d{2}) (\d{4})\)/
-        );
-        const matchResult2 = orderDate2.match(
-            /(\d{1,2}):(\d{2}):(\d{2}) (AM|PM) \((\w{3}) (\w{3}) (\d{2}) (\d{4})\)/
-        );
-
-        if (matchResult && matchResult2) {
-            // converting invalid date to a valid Date for first order to be compared
-            const [, hours, minutes, seconds, meridiem, , month, day, year] =
-                matchResult;
-            const monthIndex = [
-                "Jan",
-                "Feb",
-                "Mar",
-                "Apr",
-                "May",
-                "Jun",
-                "Jul",
-                "Aug",
-                "Sep",
-                "Oct",
-                "Nov",
-                "Dec",
-            ].indexOf(month);
-
-            const date = new Date(
-                Number(year),
-                monthIndex,
-                Number(day),
-                Number(hours) + (meridiem === "PM" ? 12 : 0),
-                Number(minutes),
-                Number(seconds)
-            );
-            const formattedDate = date.toISOString();
-
-            // converting invalid date to a valid Date for first order to be compared
-            const [, hours2, minutes2, seconds2, meridiem2, , month2, day2, year2] =
-                matchResult2;
-            const monthIndex2 = [
-                "Jan",
-                "Feb",
-                "Mar",
-                "Apr",
-                "May",
-                "Jun",
-                "Jul",
-                "Aug",
-                "Sep",
-                "Oct",
-                "Nov",
-                "Dec",
-            ].indexOf(month2);
-
-            const date2 = new Date(
-                Number(year2),
-                monthIndex2,
-                Number(day2),
-                Number(hours2) + (meridiem2 === "PM" ? 12 : 0),
-                Number(minutes2),
-                Number(seconds2)
-            );
-            const formattedDate2 = date2.toISOString();
-
-            return (
-                new Date(formattedDate2).valueOf() - new Date(formattedDate).valueOf()
-            );
-        } else {
-            console.log("Invalid time format");
-        }
-        return 0;
-    });
+    const orders = unsorted_orders.slice().sort(sortReturnedOrders);
 
     const [isVisible, setVisibility] = useState(true);
     const toggleVisibility = () => setVisibility(!isVisible);
